+++ packages/orq-rc/src/orq_ai_sdk/models/getagentop.py
@@ -303,7 +303,7 @@ class GetAgentKnowledgeBases(BaseModel):
     configuration: GetAgentKnowledgeBaseConfiguration
     r"""Defines the configuration settings which can either be for a user message or a text entry."""
 
-    id: Optional[str] = "01K76WWQKD6CCGY96SW78PA1SY"
+    id: Optional[str] = "01K7C4065DRNQYGRQ97G9XWKMZ"
     r"""The id of the resource"""