+++ packages/orq-rc/src/orq_ai_sdk/models/listagentsop.py
@@ -298,7 +298,7 @@ class ListAgentsKnowledgeBases(BaseModel):
     configuration: ListAgentsKnowledgeBaseConfiguration
     r"""Defines the configuration settings which can either be for a user message or a text entry."""
 
-    id: Optional[str] = "01K76WWQKFAE3Q84HR3XAKG8NS"
+    id: Optional[str] = "01K7C4065FQ8VDTT7945BEX1JA"
     r"""The id of the resource"""