+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-10T11:07:48.042Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-12T11:48:13.882Z")
     r"""The date and time the resource was last updated"""