+++ packages/orq-rc/src/orq_ai_sdk/models/retrievecontactop.py
@@ -103,7 +103,7 @@ class RetrieveContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-10T11:07:48.042Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-12T11:48:13.882Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")