+++ packages/orq-rc/RELEASES.md
@@ -3249,3 +3249,13 @@ Based on:
 - [python v3.14.0-rc.20] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.20] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.20 - packages/orq-rc
+
+## 2025-10-12 11:49:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.3 (2.723.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.21] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.21] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.21 - packages/orq-rc